@@ -12,18 +12,19 @@
   "license": "ISC",
   "dependencies": {
     "@babel/parser": "^7.19.4",
+    "chalk": "^4.1.2",
     "command-line-args": "^5.2.1",
     "eslint": "^8.23.0",
     "fast-xml-parser": "^4.0.9",
+    "fp-ts": "^2.16.2",
     "puppeteer": "^18.2.1",
-    "typescript": "^4.8.4",
-    "chalk": "^4.1.2"
+    "typescript": "^5.3.2"
   },
   "devDependencies": {
     "@babel/types": "^7.19.4",
     "@types/command-line-args": "^5.2.0",
     "@types/eslint": "^8.4.6",
-    "@types/node": "^18.8.2",
+    "@types/node": "^20.11.24",
     "@types/puppeteer": "^5.4.7"
   }
 }
